requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
tensorflow
streamlit
keras
ipykernel
Pillow